+++ test/dummy/app/controllers/admin_protected_controller.rb
@@ -1,0 +1,7 @@
+class AdminProtectedController < ApplicationController
+  before_action :authenticate_admin
+
+  def index
+    head :ok
+  end
+end