+++ test/dummy/app/models/admin.rb
@@ -1,0 +1,3 @@
+class Admin < ActiveRecord::Base
+  has_secure_password
+end