+++ test/dummy/config/routes.rb
@@ -1,5 +1,8 @@
 Rails.application.routes.draw do
   resources :protected_resources
   resource :current_user
+
+  resources :admin_protected
+
   mount Knock::Engine => "/knock"
 end