+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
-## Added
+## [Unreleased] - Unreleased
+### Added
+- Exception configuration option `Knock.not_found_exception_class_name`
+- Multiple entity authentication (e.g. User, Admin, etc)
 - Possibility to have permanent tokens
 - adding config options for exception class
+
+### Changed
+- Deprecated `Authenticable#authenticate` in favor of `Authenticable#authenticate_user`
+- Deprecated use of `Knock.current_user_from_token` in favor of `User.find_for_authentication`
 
 ## [1.4.2] - 2016-01-29
 ### Fixed